SqlAlchemy: fix bug in `Group` extras migration with revision `0edcdd5a30f0`
This migration added the JSONB column `extras` to the `DbGroup` model.
Since this needs to be a non-nullable column, but existing group rows
would have a null value, the migration was implemented by first adding
the column as nullable, after which all rows had their value set to the
default, before making the column non-nullable.

This approach was working when the revision was run by itself in an
isolated way, which is how the unit test is run, and so that passed.
However, as soon as it was run with multiple revisions in a sequence the
migration would fail with the message that the `db_dbgroup` table has
pending event triggers as soon as the final instruction to change the
nullability of the column was executed. The cause for this was never
found.

As an alternative, the migration is adapted to instead create the column
directly as non-nullable, but also provide a server default. This will
also avoid the exception of existing rows with a null value, just as the
original approach, but now when we remove the server default in a second
operation, we no longer hit the problem of the pending trigger events.