CI: manually install `numpy==1.19.4` to prevent release candidate
This is the usual problem that occurs when `numpy` releases a candidate
release. Even though we specify limits to not install this, `pymatgen`
installs it as a dependency in their build process where limits are
ignored and so the latest release gets installed. In this case
`numpy==1.20.0rc1` is getting installed and that dropped support for
Python 3.6 and so our builds on that version brick.

The workaround is to manually install the compatible version, in this
case `numpy==1.19.4` before installing `aiida-core`. This way `pymatgen`
will simply use that in their build process.