CI: manually install to prevent incompatible releases
When `pymatgen` gets installed through `pip` it needs to be built and
its build process will not respect any package version requirements that
`aiida-core` imposes. A common problem is that the latest `numpy`
release will get installed, including pre-releases, which often break
our tests because it drops support of certain Python versions.

To workaround this, the workflows are updated to install an explicit
version of `numpy` that we know is compatible before installing
`aiida-core` itself. In this case, `pymatgen` should simply use that
version and not install the latest.

Cherry-pick: c063609f48c494418164c3e043568628bf66c3a8