Data: add plugin type `RemoteFolderData` and `RemoteStashData`
These new data plugins are a subclass of the existing `RemoteData`. The
`RemoteFolderData` is identical to the original `RemoteData` and is
essentially replacing it. The `RemoteData` type will at some point
become unstorable and will then merely function as a parent class for
querying. The `RemoteStashData` is new and is the second remote type of
data and will be used to "stash" data from the remote working directory
of a completed calculation job for safe keeping. Since this data remains
also on the remote, it is "symbolic" just as the `RemoteFolderData`
because only a reference of its location on the remote is kept. If the
data is removed on the remote, over which AiiDA has no control, any
process that uses it as input is no longer reproducible. The symbolic
data will have to be recreated first.

Since the two types share this "symbolic" character, they need to have a
shared parent class. For backward compatiblity reasons, `RemoteData` is
chosen as that parent class which means the entry point string can
remain unchanged and the subclasses can extend the `remote` namespace.