`RemoteData`: deprecate automatic opening of a transport
The methods of the `RemoteData` class that require a transport now, for
convenience, automatically open that transport for the caller. However,
in a high-throughput context this is not scalable as each call opening
its own transport will incur too many opened connections to the target
machine.

Instead, a single transport should be opened that is then passed to the
method calls. All affected methods now have an argument `transport` that
allows to pass a transport in. If it is not passed a deprecation warning
is emitted as this usage will raise in the future as the argument will
be made required.

Note that there is one backwards-incompatible change. The `is_empty`
property had to be changed into a method in order for it to take the
transport as argument. It is not clear whether there is a scheme in
which a property can be transformed into a method in a backwards
compatible way through a deprecation pathway.