Major refactoring of the exec_commmand_wait functionality
Now, I add a exec_commmand_wait_bytes that actually does the job, and this
needs to be implemented by plugins. The two plugins implemented by AiiDA
already define instead the new function.
Also, I improved in both the API for the stdin, so that these commands
can accept also bytes ot BytesIO objects, that makes more sense in general.

I tried to keep the API (when calling them) backward-compatible, so using e.g.
a str/StringIO as stdin still works, and similarly `exec_command_wait` works
as before, but adds an optional `encoding` parameter (the default is `utf-8` to
give the same old behavior, but give more flexibility now in using different
encodings).
Tests have been added for these usecases.

However, note that the interface of plugins have changed (plugins now have to define
instead the `exec_command_wait_bytes` function instead). The change is minimal
(change the function name and avoid decoding the bytes into strings) but plugins
extending the transport functionality should take care of this (and, if they drop
the `exec_command_wait`, they should be depending on AiiDA >=1.6; otherwise they
should copy the implementation from the general Transport class if they want to remain
backward-compatible).

Also, I changed a couple of places to call directly the _bytes version, to avoid
possible issues with strange encodings on the remote computer.
Others are left in to avoid too many changes in the code, since until now noone complained
because they had a weird encoding on the supercomputer.
When such issue will arise we'll fix it, and thanks to this PR now the code is all ready
to move to treat bytes directly (or use a custom encoding, that e.g. might need to be
defined in the `verdi computer setup` or `verdi computer configure` steps, in the future).