Optimizing the parameters to have a high throughput
I could get ~100MB/s when not using compression, ~33MB/s when using it
(the second case is capped by the compression speed; the time is only
marginally larger of the one I get if I run, on the command line,
`time ssh -C localhost cat /path/to/remote/file | md5sum`;
the first one is not as good as running md5 (via SSH) on the command line
(which can go ~3x faster of a 4GB file)
but it's still acceptable for most usecases.

(Note that here we're speaking of differences measurable only when
sending hundreds of MB sent over stdout, which should not be anyway
the default way of transferring a lot of data).