Fixing the repository-writing deprecations
A few tests were writing directly to a node's repository folder, after
this got stored.
While this still works, this is now deprecated and will not work
anymore in 2.0.

Here I'm replacing all times we use `.open` with `'w'` or `'wb'` mode
with a correct call to `put_object_from_filelike`, calling it
*before* the node is stored.

In one case, the data comes from a small archive file. In this case,
I recreated the (zipped) .aiida file adding two additional (binary) files
obtained by gzipping a short string.
This was used to ensure that `inputcat` and `outputcat` work also
when binary data was requested. Actually, this is better than before,
where the actual input or output of the calculation were overwritten
and then replaced back.

This commit fixes #4721