Removing additional deprecation warnings
Typically AiiDA deprecation warnings for tests of currently
deprecated features.
I also removed a few more warnings that should not be printed,
as well as enabled the ResourceWarnings that mention if you are
e.g. opening a file and not closing it.
I fixed also the few occurrences I was aware where this was
being done - there might be more.