Fix a few bugs while removing some warnings
A couple of files were opened and not closed
Maybe this was partially related to #4307 that closed
as a "tentative" fix of #2386, but maybe a few more
open files were remaining?

Also, I'm adding a return status to most of the verdi daemon
commands, so tests will catch possible errors (they weren't being
catched in my own setup).

In addition, I'm also silencing some additional warnings that should be
silenced (the main point of this PR).
Note that while unmuting in general ResourceWarnings was good to spot
some issues (see bug described above), there are a couple more that now
issue it and are not obvious to fix (typically related to the daemon
starting some process in the background - or being started itself -
and not being closed before the test actually finished).
I think this is an acceptable compromise - maybe we'll figure out
how to selectively silence those, and keeping warnings visible will
help us figure out possible leaks in the future.