Database: add cross-reference of repository UUID to settings table
The new repository implementation, using the `disk-objectstore`
underneath, now provides a UUID for each repository. Currently, each
database can only be configured to work with a single repository. By
writing the UUID of the repository into the database, it will become
possible to enable a consistency check of the repository and database
that are configured for a particular profile. This will prevent
accidental misconfigurations where the wrong repository is coupled to a
particular database.

The repository UUID is generated automatically by the `disk-objectstore`
library when the container is created and it provides an interface to
retrieve it. This value is written to the database in the `verdi setup`
command when a new profile is created. If the database already has the
repository UUID setting defined, it will be cross-referenced with the
one of the repository to make sure it is compatible. This case is to
facilitate the creation of a new profile for an existing repository and
database. However, if the UUIDs do not match, the setup command fails.

For databases that were created before this commit, the database
migration that performed the migration of the legacy file repository
includes a statement that will insert the UUID of the new object store
container once is has been created.