Profile: check database and repository compatibility on load
Since the migration to the new repository implementation, each file
repository has its own UUID. This UUID is now written to the settings
table of the database that is associated to it. This allows to check
that the repository and database that are configured for a profile are
compatible.

The check is added to the `Manager._load_backend` method, as this is the
central point where the database is loaded for the currently loaded
profile. We need to place the check here since in order to retrieve the
repository UUID from the database, the corresponding backend needs to be
loaded first.

If the UUID of the repository and the one stored in the database are
found to be different, a warning is emitted instructing the user to make
sure the repository and database are correctly configured. Since this is
new functionality and its stability is not known, a warning is chosen
instead of an exception in order to prevent AiiDA becoming unusable in
the case of an unintentional bug in the compatibility checking. In the
future, when the check has been proven to be stable and reliable, the
warning can be changed into an exception.