CLI: Use DB `proxy_command` for `verdi calcjob gotocomputer` (#4761)
Currently the `verdi calcjob gocomputer` command only works for
computers accessed via a proxy in case:

* the `Host` is set up in the `.ssh/config` file. 
* the `Host` is the same as the `HostName`, also configured as the host
  name for the AiiDA computer.

If both conditions are met, the `verdi calcjob gocomputer` command
generates an `ssh` command (via `SshTransport.gotocomputer_command()`)
that can rely on the configuration in the `.ssh/config` to use the proxy 
correctly, because the generated command is executed via `os.system()`
and hence it parses the `.ssh/config` file. However, since typically users set
an alias `Host` instead of using the full `HostName`, this will often not work.

Here we add the `proxy_command` that has been configured for the
`Computer` in the database to the ssh command string that is created for 
`verdi calcjob gotocomputer`. This fixes the `verdi calcjob gotocomputer`
command for computers which are accessed by a proxy in case the above
conditions are not met and the user has configured the `proxy_command`
for the `Computer`.