Fix `aiida.cmdline.utils.decorators.with_dbenv` always loading the db
The premise of the method `with_dbenv` is that it will load the profile
and corresponding database environment if not already loaded. However,
due to a bug in `load_backend_if_not_loaded` it was actually *always*
loading the environment for every call, even if the database env was
already loaded, leading to very expensive calls for nothing.