`NodeLinksManager`: add support for nested namespaces
The `NodeLinksManager` is used by the ``inputs`` and ``outputs``
attributes of the ``ProcessNode`` class to allow users to quickly access
input and output nodes by their link label. Nested namespaces in link
labels are converted to double underscores before it is stored in the
database because it can only store a flat string. This is an
implementation detail, however, and the user should not have to know
about it and should be able to use the nested namespace. However, up
till now, one had to pass the link label as stored in the database, i.e.

    node.inputs.nested__sub__namespace

After this commit, it is now possible to use the more intuitive

    node.inputs.nested.sub.namespace

For backwards compatibility, the old flat link is still supported but
will emit a deprecation warning.