CLI: fix loading time of `verdi`
The `from disk_objectstore import Container` import has a significant
loading time. The `aiida.manage.configuration.profile` module imported
it at the top level, and since this module is loaded when `verdi` is
called, the load time of `verdi` was significantly increased. This would
have a detrimental effect on the tab completion speed.

The work around is to only import the module within the methods that use
it. Ideally, the loading of this library would not be so costly.