FIX: respect nested output namespaces in `Process.exposed_outputs` (#4863)
The `exposed_outputs` method was detecting nested namespaces in the
dictionary of outputs of a node by using the namespace separator
character of the port namespace class. However, this character, which is
the `.`, gets converted to `__` when the link is stored in the database.
This transformation is necessary since the `.` is a reserved character
to enable attribute based dereferencing, e.g.

    node.outputs.some_output

Since link labels can contain underscores, we use a double underscore
which, together with the guarantee that link labels don't have leading
or terminating underscores, can uniquely determine the namespaces in any
given link label.

The solution is to not manually reconstruct the namespaces in the output
dictionary but simply use the `get_outgoing().nested()` method which
does it for us.